requirements.txt
git+https://github.com/adsabs/ADSMicroserviceUtils.git@v1.1.9
flask-headers
git+https://github.com/adsabs/flask-limiter.git@fb1bbd3e4894c4298798601121dbf9782a6bafe8#egg=flask-limiter
flask-cors
flask-sslify
psycopg2
Flask-login
Flask-Security
Flask-Menu
Flask-Script
Flask-Breadcrumbs
oauthlib
git+https://github.com/lepture/flask-oauthlib.git@e092a5f46ff6c78d313a6eb2f89ee746fb190dfb#egg=Flask-OAuthlib
Flask-WTF
WTForms-Components
WTForms-Alchemy
Flask-Registry
Flask-Email
Flask-Mail
sqlalchemy-utils
alembic
redis
flask-consulate
email-validator
jsondiff
sqlalchemy-citext
requests>=2.26